% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounts-data.R
\docType{data}
\name{accounts}
\alias{accounts}
\title{Accounts Receivable}
\format{
A data frame with 20 rows and 3 variables.
\describe{
  \item{account}{account number (between 1 - 87)}
  \item{bookValue}{booked value of the account}
  \item{auditValue}{audited (true) value of the account}
}
}
\usage{
data(accounts)
}
\description{
Audit sample obtained from a population of N = 87 accounts receivable,
totaling $612,824 in book value (Higgins and Nandram, 2009; Lohr, 2021).
}
\examples{
data(accounts)
}
\references{
Higgins, H. N., & Nandram, B. (2009). Monetary unit sampling:
  Improving estimation of the total audit error
  \emph{Advances in Accounting}, 25(2), 174-182.
  \doi{10.1016/j.adiac.2009.06.001}

Lohr, S. L. (2021). \emph{Sampling: Design and Analysis}. CRC
  press.
}
\keyword{datasets}
