% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allowances-data.R
\docType{data}
\name{allowances}
\alias{allowances}
\title{Legitimacy Audit}
\format{
A data frame with 4189 rows and 5 variables.
\describe{
  \item{item}{unique record identification number.}
  \item{branch}{branch number.}
  \item{bookValue}{book value in US dollars.}
  \item{auditValue}{true value in US dollars.}
  \item{times}{sample indicator.}
}
}
\usage{
data(allowances)
}
\description{
Fictional population from a legitimacy audit, containing 4189 observations identification numbers and stratum identifiers, book values, and audit values. The audit values are added for illustrative purposes, as these would need to be assessed by the auditor in the execution stage of the audit.
}
\examples{
data(allowances)
}
\keyword{datasets}
