% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditBF.R
\name{auditBF}
\alias{auditBF}
\title{Bayes Factors for Audit Sampling}
\usage{
auditBF(x, n, materiality, likelihood = 'poisson', BF10 = TRUE,
        log = FALSE, N.units = NULL, alpha = NULL, beta = NULL)
}
\arguments{
\item{x}{a number larger than zero specifying the observed proportional error (i.e., sum of taints) in the sample.}

\item{n}{an integer larger than 0 specifying the number of items in the sample (i.e., the sample size).}

\item{materiality}{a numeric value between 0 and 1 specifying the performance materiality (i.e., the maximum upper limit) as a fraction of the total population size. Can be \code{NULL} for some methods.}

\item{likelihood}{a character specifying the likelihood assumed when updating the prior distribution. This can be either \code{binomial} for the binomial likelihood and beta prior distribution, \code{poisson} for the Poisson likelihood and gamma prior distribution, or \code{hypergeometric} for the hypergeometric likelihood and beta-binomial prior distribution. See the details section for more information about the available likelihoods.}

\item{BF10}{logical; if \code{TRUE}, the Bayes factor computed is in favor or tolerable misstatement. If \code{FALSE}, the Bayes factor is in favor of intolerable misstatement.}

\item{log}{logical; if \code{TRUE}, the Bayes factor is given as log(bf).}

\item{N.units}{an integer larger than 0 specifying the total number of sampling units in the population. Only required when \code{likelihood = 'hypergeometric'}.}

\item{alpha}{a numeric value specifying the \eqn{\alpha} parameter of the prior distribution. If specified, overrides the default parameters from the impartial prior distribution.}

\item{beta}{a numeric value specifying the \eqn{\beta} parameter of the prior distribution. If specified, overrides the default parameters from the impartial prior distribution.}
}
\value{
A numeric value for the Bayes factor.
}
\description{
This function computes Bayes factors for audit sampling from summary statistics of an audit sample. By default, the Bayes factor is computed using an impartial prior distribution on the misstatement (Derks et al., 2021). However, the arguments \code{alpha} and \code{beta} can be used to specify an alternative prior distribution.

For more details on how to use this function, see the package vignette:
\code{vignette('jfa', package = 'jfa')}
}
\details{
The Bayes Factor \eqn{BF_{-+}} quantifies how much more likely the data are to be observed under \eqn{H_{-}: \theta < \theta_{max}} than under \eqn{H_{+}: \theta > \theta_{max}}. Therefore, \eqn{BF_{-+}} can be interpreted as the relative support in the observed data for \eqn{H_{-}} versus \eqn{H_{+}}. If \eqn{BF_{-+}} is 1, there is no preference for either \eqn{H_{-}} or \eqn{H_{+}}. If \eqn{BF_{-+}} is larger than 1, \eqn{H_{-}} is preferred. If \eqn{BF_{-+}} is between 0 and 1, \eqn{H_{+}} is preferred.

This section elaborates on the available likelihoods and corresponding prior distributions for the \code{likelihood} argument.

\itemize{
 \item{\code{poisson}:          The Poisson likelihood is often used as a likelihood for monetary unit sampling (MUS). The likelihood function is defined as: \deqn{p(x) = \frac{\lambda^x e^{-\lambda}}{x!}} The conjugate \emph{gamma(\eqn{\alpha, \beta})} prior has probability density function: \deqn{f(x; \alpha, \beta) = \frac{\beta^\alpha x^{\alpha - 1} e^{-\beta x}}{\Gamma(\alpha)}}}
 \item{\code{binomial}:         The binomial likelihood is often used as a likelihood for attributes sampling \emph{with} replacement. The likelihood function is defined as: \deqn{p(x) = {n \choose k} p^k (1 - p)^{n - k}} The conjugate \emph{beta(\eqn{\alpha, \beta})} prior has probability density function: \deqn{f(x; \alpha, \beta) = \frac{1}{B(\alpha, \beta)} x^{\alpha - 1} (1 - x)^{\beta - 1}}}
 \item{\code{hypergeometric}:   The hypergeometric likelihood is used as a likelihood for sampling \emph{without} replacement. The likelihood function is defined as: \deqn{p(x = k) = \frac{{K \choose k} {N - K \choose n - k}}{{N \choose n}}} The conjugate \emph{beta-binomial(\eqn{\alpha, \beta})} prior (Dyer and Pierce, 1993) has probability density function: \deqn{f(k | n, \alpha, \beta) = {n \choose k} \frac{B(k + \alpha, n - k + \beta)}{B(\alpha, \beta)}} }
}
}
\examples{
 
# Compute a default Bayes factor from an impartial gamma prior
auditBF(x = 1, n = 50, materiality = 0.05)

# Compute a Bayes factor from a noninformative gamma prior
auditBF(x = 1, n = 50, materiality = 0.05, alpha = 1, beta = 1)

}
\references{
Derks, K., de Swart, J., van Batenburg, P., Wagenmakers, E.-J., & Wetzels, R. (2021). Priors in a Bayesian audit: How integration of existing information into the prior distribution can improve audit transparency and efficiency. \emph{International Journal of Auditing}, 1-16.
}
\seealso{
\code{\link{auditPrior}} \code{\link{planning}} \code{\link{selection}} \code{\link{evaluation}} \code{\link{report}}
}
\author{
Koen Derks, \email{k.derks@nyenrode.nl}
}
\keyword{Bayes}
\keyword{audit}
\keyword{distribution}
\keyword{factor}
\keyword{prior}
