% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-data.R
\docType{data}
\name{benchmark}
\alias{benchmark}
\title{Benchmark Analysis of Sales Versus Cost of Sales}
\format{
A data frame with 100 rows and 2 variables.
\describe{
  \item{sales}{book value in US dollars ($100,187,432--$398,280,933).}
  \item{costofsales}{true value in US dollars ($71,193,639--$309,475,784).}
}
}
\usage{
data(benchmark)
}
\description{
Fictional data from a benchmark analysis comparing industry sales versus
the industry cost of sales.
}
\examples{
data(benchmark)
}
\references{
Derks, K., de Swart, J., van Batenburg, P., Wagenmakers, E.-J.,
  & Wetzels, R. (2021). Priors in a Bayesian audit: How integration of
  existing information into the prior distribution can improve audit
  transparency and efficiency. \emph{International Journal of Auditing},
  25(3), 621-636. \doi{10.1111/ijau.12240}
}
\keyword{datasets}
