% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.n.beta.R
\name{calc.n.beta}
\alias{calc.n.beta}
\title{Audit Sample Size for any Sample}
\usage{
calc.n.beta(expected.mistakes, materiality, mean.taint = 1, confidence = 0.95)
}
\arguments{
\item{expected.mistakes}{An integer representing the number of expected mistakes
in the sample.}

\item{materiality}{A value representing the materiality of the audit in percentages.}

\item{mean.taint}{A value representing the expected mean taint in the sample, defaults to 1.}

\item{confidence}{The amount of confidence desired from the bound
(on a scale from 0 to 1), defaults to 95\% confidence.}
}
\value{
A value indicating the required sample size for the audit.
}
\description{
This function calculates the required sample size for an audit, based on
the Beta distribution. The Beta distribution is not bound to a population
size, and can be used in any case. calc.n.beta improves on calc.n.binomial by
presenting the possibility that the mistakes are not complete mistakes.
}
\section{Details}{
 EMPTY FOR NOW
}

\examples{
# Calculate the required sample size for a materiality of 5\\\% when one mistake
# is expected to be found in the sample, which is expected to have a taint of 0.5.
calc.n.binomial(expected.mistakes = 1,
                materiality = 0.05,
                mean.taint = 0.5,
                confidence = 0.95)

}
\references{

}
\seealso{

}
\author{
Koen Derks, \email{k.derks@nyenrode.nl}
}
\keyword{sample}
\keyword{size}
