% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.n.binomial.R
\name{calc.n.binomial}
\alias{calc.n.binomial}
\title{Audit Sample Size for Large Samples}
\usage{
calc.n.binomial(expected.mistakes, materiality, confidence = 0.95)
}
\arguments{
\item{expected.mistakes}{An integer representing the number of expected mistakes
in the sample.}

\item{materiality}{A value representing the materiality of the audit in percentages.}

\item{confidence}{The amount of confidence desired from the bound
(on a scale from 0 to 1), defaults to 95\% confidence.}
}
\value{
A value indicating the required sample size for the audit.
}
\description{
This function calculates the required sample size for an audit, based on
the Binomial distribution. The Binomial distribution can be used if the
expected required sample is less than 10\% of the population.
\deqn{\frac{n}{N} < 0.10}{n / N < 10\%}
}
\section{Details}{
 EMPTY FOR NOW
}

\examples{
# Calculate the required sample size for a materiality of 5\\\% when one mistake
# is expected to be found in the sample.
calc.n.binomial(expected.mistakes = 1,
                materiality = 0.05,
                confidence = 0.95)

}
\references{

}
\seealso{

}
\author{
Koen Derks, \email{k.derks@nyenrode.nl}
}
\keyword{sample}
\keyword{size}
