% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrier-data.R
\docType{data}
\name{carrier}
\alias{carrier}
\title{Carrier Company Financial Statements}
\format{
A data frame with 202 rows and 12 variables.
\describe{
  \item{description}{description of the ledger item.}
  \item{entity1 - entity 10}{recorded values for entity 1 - 10, in US dollars.}
  \item{total}{total value, in US dollars.}
}
}
\source{
\url{https://towardsdatascience.com/data-driven-audit-1-automated-sampling-using-python-52e83347add5}
}
\usage{
data(carrier)
}
\description{
Fictional data from a carrier company in Europe, containing 202 ledger items across 10 company entities.
}
\examples{
data(carrier)
}
\keyword{datasets}
