% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compas-data.R
\docType{data}
\name{compas}
\alias{compas}
\title{COMPAS Recidivism Prediction}
\format{
A data frame with 100 rows and 2 variables.
\describe{
  \item{TwoYrRecidivism}{yes/no for recidivism or no recidivism.}
  \item{AgeAboveFoutryFive}{yes/no for age above 45 years or not}
   \item{AgeBelowTwentyFive}{yes/no for age below 25 years or not}
   \item{Gender}{female/male for gender}
   \item{Misdemeanor}{yes/no for having recorded misdemeanor(s) or not}
   \item{Ethnicity}{Caucasian, African American, Asian, Hispanic, Native American or Other}
   \item{Predicted}{yes/no, predicted values for recidivism}
}
}
\usage{
data(compas)
}
\description{
This data was used to predict recidivism (whether a criminal will reoffend or
not) in the USA.
}
\examples{
data(compas)
}
\references{
\url{https://www.kaggle.com/danofer/compass}
            \url{https://cran.r-project.org/package=fairness}
}
\keyword{datasets}
