% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidenceBound.R
\name{confidenceBound}
\alias{confidenceBound}
\title{Audit confidence / credible bounds}
\usage{
confidenceBound(sample = NULL, bookValues = NULL, auditValues = NULL, confidence = 0.95,
                       dataType = "sample", sampleSize = NULL, sumErrors = NULL,
                       method = "binomial", materiality = NULL)
}
\arguments{
\item{sample}{a data frame containing a column of book values and a column of audit values.}

\item{bookValues}{the column name for the book values in the sample.}

\item{auditValues}{the column name for the audit (true) values in the sample.}

\item{confidence}{the required confidence for the bound.}

\item{dataType}{can be either "sample" for data input, or "sumstats" for input in the form of summary statistics.}

\item{sampleSize}{the number of observations in the sample. Only used when dataType = "sumstats".}

\item{sumErrors}{the sum of the errors found in the sample. Only used when dataType = "sumstats".}

\item{method}{can be either one of "binomial", "hypergeometric", "poisson", "stringer".}

\item{materiality}{if specified, the function also returns the conclusion of the analysis with respect to the materiality.}
}
\value{
A list containing the confidence bound for the audit.
}
\description{
This function takes summary statistics about an evaluated audit sample and calculates
a confidence bound accordint to a specified method.
}
\examples{

# Using the binomial distribution, calculates the upper confidence bound for a materiality of 5\% 
# when 3 mistakes are found in a sample of 234.

# Frequentist planning (n = 234)
jfaRes <- jfa::sampleSize(materiality = 0.05, confidence = 0.95,
                          expectedError = 0.025, likelihood = "binomial")

# Six errors are allowed in the sample. Three are found.

confidenceBound(sampleSize = jfaRes$sampleSize, sumErrors = 3, method = "binomial")

}
\references{

}
\seealso{

}
\author{
Koen Derks, \email{k.derks@nyenrode.nl}
}
\keyword{bound}
\keyword{confidence}
