% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox.snell.R
\name{cox.snell}
\alias{cox.snell}
\title{Cox and Snell bound}
\usage{
cox.snell(bookValues, auditValues, priorPi = 0.10, priorMu = 0.40, priorA = 1, priorB = 6,
           confidence = 0.95)
}
\arguments{
\item{bookValues}{A vector of book values from sample.}

\item{auditValues}{A vector of corresponding audit values from the sample.}

\item{priorPi}{A value between 0 and 1 defining the prior expected error rate.}

\item{priorMu}{A value between 0 and 1 defining the prior expected mean taint.}

\item{priorA}{A value between 0 and 1 defining the prior expected successes.}

\item{priorB}{A value between 0 and 1 defining the prior expected failures.}

\item{confidence}{The amount of confidence desired from the bound
(on a scale from 0 to 1), defaults to 95\% confidence.}
}
\value{
An estimate of the mean taint per dollar unit in the population.
}
\description{
Calculates the Cox and Snell confidence bound for the maximum error in an
audit population according to the methodology described by Cox & Snell (1979).
}
\section{Details}{
 EMPTY FOR NOW
}

\examples{
# Create an imaginary data set
bookValues   <- rgamma(n = 2400, shape = 1, rate = 0.001)
error.rate   <- 0.1
error        <- sample(0:1, 2400, TRUE, c(1-error.rate, error.rate))
taint        <- rchisq(n = 2400, df = 1) / 10
auditValues  <- bookValues - (error * taint * bookValues)
frame        <- data.frame( bookValues = round(bookValues,2),
                            auditValues = round(auditValues,2))
# Draw a sample
samp.probs   <- frame$bookValues/sum(frame$bookValues)
sample.no    <- sample(1:nrow(frame), 100, FALSE, samp.probs)
sample       <- frame[sample.no, ]
# Calculate bound
cox.snell(bookValues,
          auditValues,
          confidence = 0.95)

}
\references{
Dworin, L., & Grimlund, R. A. (1986). Dollar-unit sampling: A
comparison of the quasi-Bayesian and moment bounds. Accounting Review, 36-57.
}
\seealso{

}
\author{
Koen Derks, \email{k.derks@nyenrode.nl}
}
\keyword{bound}
