% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jfa-convolution.R
\name{jfa-convolution}
\alias{jfa-convolution}
\alias{+.jfaPrior}
\alias{+.jfaPosterior}
\alias{*.jfaPrior}
\alias{*.jfaPosterior}
\title{Prior convolution utility functions}
\usage{
\method{+}{jfaPrior}(p1, p2)

\method{+}{jfaPosterior}(p1, p2)

\method{*}{jfaPrior}(p1, p2)

\method{*}{jfaPosterior}(p1, p2)
}
\arguments{
\item{p1}{an object of class \code{jfaPrior} or class \code{jfaPosterior}.}

\item{p2}{an object of class \code{jfaPrior} or class \code{jfaPosterior}.}
}
\value{
An object of class \code{jfaPrior} or class \code{jfaPosterior}
constructed with method \code{convolution}.
}
\description{
Methods defined for convolutions of \code{jfaPrior} objects.
}
