% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified.moment.R
\name{modified.moment}
\alias{modified.moment}
\title{Modified Moment bound}
\usage{
modified.moment(bookValues, auditValues, pop.type = "inventory", confidence = 0.95)
}
\arguments{
\item{bookValues}{A vector of book values from sample.}

\item{auditValues}{A vector of corresponding audit values from the sample.}

\item{pop.type}{A character defining the type of population audited.
\emph{inventory} for inventory populations. \emph{accounts} for populations
of accounts receivable.}

\item{confidence}{The amount of confidence desired from the bound
(on a scale from 0 to 1), defaults to 95\% confidence.}
}
\value{
An estimate of the mean taint per dollar unit in the population.
}
\description{
Calculates the Modified Moment confidence bound for the maximum error in an
audit population according to the methodology described by Dworin &
Grimlund (1986).
}
\section{Details}{
 EMPTY FOR NOW
}

\examples{
# Create an imaginary data set
bookValues   <- rgamma(n = 2400, shape = 1, rate = 0.001)
error.rate   <- 0.1
error        <- sample(0:1, 2400, TRUE, c(1-error.rate, error.rate))
taint        <- rchisq(n = 2400, df = 1) / 10
auditValues  <- bookValues - (error * taint * bookValues)
frame        <- data.frame( bookValues = round(bookValues,2),
                            auditValues = round(auditValues,2))
# Draw a sample
samp.probs   <- frame$bookValues/sum(frame$bookValues)
sample.no    <- sample(1:nrow(frame), 100, FALSE, samp.probs)
sample       <- frame[sample.no, ]
# Calculate bound
modified.moment(bookValues = sample$bookValues,
                auditValues = sample$auditValues,
                pop.type = "inventory",
                confidence = 0.95)

}
\references{
Dworin, L., & Grimlund, R. A. (1986). Dollar-unit sampling: A
comparison of the quasi-Bayesian and moment bounds. Accounting Review, 36-57.
}
\seealso{

}
\author{
Koen Derks, \email{k.derks@nyenrode.nl}
}
\keyword{bound}
