% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.sampling.R
\name{random.sampling}
\alias{random.sampling}
\title{Simple Random Sampling for Audit Populations}
\usage{
random.sampling(data, bookCol, sample.size)
}
\arguments{
\item{data}{A data frame containing at least the book values of the audit
population.}

\item{bookCol}{A character indicating the name of the column used for the book
values.}

\item{sample.size}{An integer representing the number of line items to sample.}
}
\value{
A data frame containing the sample values.
}
\description{
This function takes as input a data frame containing an audit population,
which needs at least a column indicating the book values of the line items.
}
\section{Details}{
 EMPTY FOR NOW
}

\examples{
# Create an imaginary data set
bookValues   <- rgamma(n = 2400, shape = 1, rate = 0.001)
error.rate   <- 0.1
error        <- sample(0:1, 2400, TRUE, c(1-error.rate, error.rate))
taint        <- rchisq(n = 2400, df = 1) / 10
auditValues  <- bookValues - (error * taint * bookValues)
frame        <- data.frame( bookValues = round(bookValues,2),
                            auditValues = round(auditValues,2))
# Sample from the population
random.sampling(data = frame,
                bookCol = "bookValues",
                sample.size = 100)

}
\references{

}
\seealso{

}
\author{
Koen Derks, \email{k.derks@nyenrode.nl}
}
\keyword{sampling}
