% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rohrbach.bound.R
\name{rohrbach.bound}
\alias{rohrbach.bound}
\title{Rohrbach's Augmented Variance Estimator Bound}
\usage{
rohrbach.bound(bookValues, auditValues, population.size, delta = 2.7, confidence = 0.95)
}
\arguments{
\item{bookValues}{A vector of book values from sample.}

\item{auditValues}{A vector of corresponding audit values from the sample.}

\item{population.size}{An integer representing the size of the population.}

\item{delta}{A value representing the adjustment factor for the variance.
Rohrbach (1993) argued that the smallest value of delta that yields nominal
coverage equals 2.7.}

\item{confidence}{The amount of confidence desired from the bound
(on a scale from 0 to 1), defaults to 95\% confidence.}
}
\value{
An estimate of the mean taint per dollar unit in the population.
}
\description{
Calculates Rohrbach's (1993) augmented variance estimator confidence bound for
the maximum error in an audit population.
}
\section{Details}{
 EMPTY FOR NOW
}

\examples{
# Create an imaginary data set
bookValues   <- rgamma(n = 2400, shape = 1, rate = 0.001)
error.rate   <- 0.1
error        <- sample(0:1, 2400, TRUE, c(1-error.rate, error.rate))
taint        <- rchisq(n = 2400, df = 1) / 10
auditValues  <- bookValues - (error * taint * bookValues)
frame        <- data.frame( bookValues = round(bookValues,2),
                            auditValues = round(auditValues,2))
# Draw a sample
samp.probs   <- frame$bookValues/sum(frame$bookValues)
sample.no    <- sample(1:nrow(frame), 100, FALSE, samp.probs)
sample       <- frame[sample.no, ]
# Calculate bound
rohrbach.bound(bookValues = sample$bookValues,
               auditValues = sample$auditValues,
               population.size = 2400,
               confidence = 0.95)

}
\references{
Rohrbach, K. J. (1993). Variance augmentation to achieve nominal
coverage probability in sampling from audit populations. Auditing, 12(2), 79.
}
\seealso{

}
\author{
Koen Derks, \email{k.derks@nyenrode.nl}
}
\keyword{bound}
