% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleSize.R
\name{sampleSize}
\alias{sampleSize}
\title{Audit sample sizes}
\usage{
sampleSize(materiality = NULL, confidence = 0.95, expectedError = 0, distribution = "poisson", 
                  errorType = "percentage", N = NULL, maxSize = 5000)
}
\arguments{
\item{materiality}{A value representing the materiality of the audit in percentages.}

\item{confidence}{The amount of confidence desired from the bound
(on a scale from 0 to 1), defaults to 95\% confidence.}

\item{expectedError}{A percentage representing the number of expected mistakes in the sample.}

\item{likelihood}{can be one of "binomial", "poisson", or "hypergeometric"}

\item{errorType}{Whether the errors are a percentage of the total or an absolute value}

\item{N}{The population size (required for hypergeometric calculations)}

\item{maxSize}{The maximum sample size that is considered for calculations (for efficiency). 
For low materialities, increasing this parameter may be wise.}

\item{prior}{if TRUE, add a prior distribution to be updated by the likelihood. Chooses a beta distribution
for the binomial likelihood, a gamma distribution for the poisson likelihood, and a beta-binomial distribution for the 
hypergeometric likelihood.}

\item{priorK}{The prior parameter alpha (errors in the assumed prior sample)}

\item{priorN}{The prior parameter beta (sample size of assumed prior sample)}
}
\value{
A list containing the required sample size for the audit.
}
\description{
This function calculates the required sample size for an audit, based on
the poisson, binomial, or hypergeometric likelihood. A prior can be specified
to perform Bayesian planning.
}
\examples{
# Using the binomial distribution, calculates the required sample size for a materiality of 5\\% 
# when 2.5\% mistakes are expected to be found in the sample.

# Frequentist
jfa::sampleSize(materiality = 0.05, confidence = 0.95, expectedError = 0.025, 
likelihood = "binomial")

# Bayesian (uninformative beta(1, 1) prior)
jfa::sampleSize(materiality = 0.05, confidence = 0.95, expectedError = 0.025, 
likelihood = "binomial", prior = TRUE, priorK = 0, priorN = 0)

# Bayesian (informative prior of 10 assumed correct observatiosn)
jfa::sampleSize(materiality = 0.05, confidence = 0.95, expectedError = 0.025, 
likelihood = "binomial", prior = TRUE, priorK = 0, priorN = 10)

}
\references{

}
\seealso{

}
\author{
Koen Derks, \email{k.derks@nyenrode.nl}
}
\keyword{sample}
\keyword{size}
