% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sampling}
\alias{sampling}
\title{Audit sampling}
\usage{
sampling(population, sampleSize, bookValues = NULL, 
                algorithm = "random", units = "record", intervalStartingPoint = 1,
                ordered = TRUE, ascending = TRUE, withReplacement = FALSE, seed = 1)
}
\arguments{
\item{population}{a data frame containing the population the auditor wishes to sample from.}

\item{sampleSize}{the number of observations that need to be selected from the population.}

\item{bookValues}{the name of the column containing the book values (as in the population data).}

\item{algorithm}{can be either one of \emph{random} (default) for random sampling, \emph{cell} for cell sampling, or \emph{interval} for fixed interval sampling.}

\item{units}{can be either \emph{records} for record (default) sampling, or \emph{mus} for monetary unit sampling.}

\item{intervalStartingPoint}{The starting point in the interval (used only in fixed interval sampling)}

\item{ordered}{if TRUE (default), the population is first ordered according to the value of their book values.}

\item{ascending}{if TRUE (default), order the population in ascending order.}

\item{withReplacement}{whether sampling should be performed with replacement. Defaults to FALSE.}

\item{seed}{seed to reproduce results. Default is 1.}
}
\value{
An object of class \emph{jfaSampling}
}
\description{
This function takes a data frame and performs sampling according to one of three algorithms: random sampling, cell sampling, or fixed interval sampling in combination with either record sampling or monetary unit sampling.
}
\examples{

library(jfa)

# Generate some audit data (N = 1000).
population <- data.frame(ID = sample(1000:100000, size = 1000, replace = FALSE), 
                         bookValue = runif(n = 1000, min = 100, max = 500))

# Calculate the sample size according to the binomial distribution with zero errors
n <- planning(materiality = 0.05, confidence = 0.95, expectedError = 0, 
             likelihood = "binomial")$sampleSize

# Draw sample using random record sampling
sampling(population = population, sampleSize = n, algorithm = "random", 
         units = "records", seed = 1)

# Draw sample using random monetary unit sampling
sampling(population = population, sampleSize = n, algorithm = "random", 
         units = "mus", bookValues = "bookValue", seed = 1)

}
\references{

}
\seealso{

}
\author{
Koen Derks, \email{k.derks@nyenrode.nl}
}
\keyword{sampling}
