% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitizer-data.R
\docType{data}
\name{sanitizer}
\alias{sanitizer}
\title{Factory Workers' use of Hand Sanitizer}
\format{
A data frame with 1600 rows and 1 variable.
}
\usage{
data(sanitizer)
}
\description{
Data from a study on factory workers' use of hand sanitizer. Sanitizer use was measured to a 100th of a gram.
}
\examples{
data(sanitizer)
}
\references{
[Retracted] Li, M., Sun, Y., & Chen, H. (2019). The decoy effect as a nudge: Boosting hand hygiene with a worse option. Psychological Science, 30, 139–149.
}
\keyword{datasets}
