% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinoForest-data.R
\docType{data}
\name{sinoForest}
\alias{sinoForest}
\title{Financial Statemens of Sino Forest Corporation's 2010 Report}
\format{
A data frame with 772 rows and 1 variable.
}
\usage{
data(sinoForest)
}
\description{
Financial Statemens numbers of Sino Forest Corporation's 2010 Report.
}
\examples{
data(sinoForest)
}
\references{
Nigrini, M. J. (2012). Benford's Law: Application for Forensic Accounting, Auditing and Fraud Detection. Wiley and Sons: New Jersey.
}
\keyword{datasets}
