% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringer.modified.R
\name{stringer.modified}
\alias{stringer.modified}
\title{Modified Stringer bound}
\usage{
stringer.modified(bookValues, auditValues, confidence = 0.95)
}
\arguments{
\item{bookValues}{A vector of book values from sample.}

\item{auditValues}{A vector of corresponding audit values from the sample.}

\item{confidence}{The amount of confidence desired from the bound
(on a scale from 0 to 1), defaults to 95\% confidence.}
}
\value{
An estimate of the mean taint per dollar unit in the population
}
\description{
Calculates the Modified Stringer confidence bound for the maximum error in an
audit population with the Pap and van Zuijlen (1992) adjustment.
}
\section{Details}{
 EMPTY FOR NOW
}

\examples{
# Create an imaginary data set
bookValues   <- rgamma(n = 2400, shape = 1, rate = 0.001)
error.rate   <- 0.1
error        <- sample(0:1, 2400, TRUE, c(1-error.rate, error.rate))
taint        <- rchisq(n = 2400, df = 1) / 10
auditValues  <- bookValues - (error * taint * bookValues)
frame        <- data.frame( bookValues = round(bookValues,2),
                            auditValues = round(auditValues,2))
# Draw a sample
samp.probs   <- frame$bookValues/sum(frame$bookValues)
sample.no    <- sample(1:nrow(frame), 100, FALSE, samp.probs)
sample       <- frame[sample.no, ]
# Calculate bound
stringer.modified(bookValues = sample$bookValues,
                  auditValues = sample$auditValues,
                  confidence = 0.95)

}
\references{
Pap, G., & van Zuijlen, M. C. (1996). On the asymptotic behaviour
of the Stringer bound 1. Statistica Neerlandica, 50(3), 367-389.

Stringer, K. W. (1963). Practical aspects of statistical sampling
in auditing. In Proceedings of the Business and Economic Statistics Section
(pp. 405-411). American Statistical Association.
}
\seealso{
\code{\link{stringer.bound}} \code{\link{stringer.bickel}}
\code{\link{stringer.meikle}} \code{\link{stringer.lta}}
}
\author{
Koen Derks, \email{k.derks@nyenrode.nl}
}
\keyword{bound}
