% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildIt-data.R, R/retailer-data.R
\docType{data}
\name{BuildIt}
\alias{BuildIt}
\title{BuildIt Construction Financial Statements}
\format{
A data frame with 3500 rows and 3 variables.
\describe{
  \item{ID}{unique record identification number.}
  \item{bookValue}{book value in US dollars ($14.47--$2,224.40).}
  \item{auditValue}{true value in US dollars ($14.47--$2,224.40).}
}

A data frame with 20 rows and 5 variables.
\describe{
  \item{i}{branch number.}
  \item{N}{total number of items in each branch.}
  \item{n}{number of items in sample per branch.}
  \item{x}{number of errors in sample per branch.}
}
}
\usage{
data(BuildIt)

data(retailer)
}
\description{
Fictional data from a construction company in the United States, containing 3500 observations identification numbers, book values, and audit values. The audit values are added for illustrative purposes, as these would need to be assessed by the auditor in the execution stage of the audit.

Sample outcomes per stratum from a retail company consisting of 20 branches.
}
\examples{
data(BuildIt)
data(retailer)
}
\references{
Derks, K., de Swart, J., Wagenmakers, E.-J., Wille, J., & Wetzels, R. (2019). JASP for audit: Bayesian tools for the auditing practice.

Derks, K., de Swart, J., & Wetzels, R. (2022). Een Bayesiaanse blik op gestratificeerde steekproeven heeft voordelen voor de auditor. \emph{Maandblad voor Accountancy en Bedrijfseconomie}, 96(1/2), 37-46. https://doi.org/10.5117/mab.96.78836
}
\keyword{datasets}
