% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrier-data.R
\docType{data}
\name{carrier}
\alias{carrier}
\title{Carrier Company Financial Statements}
\format{
A data frame with 202 rows and 12 variables.
\describe{
  \item{description}{description of the ledger item.}
  \item{entity1}{recorded values for entity 1, in US dollars.}
  \item{entity2}{recorded values for entity 2, in US dollars.}
  \item{entity3}{recorded values for entity 3, in US dollars.}
  \item{entity4}{recorded values for entity 4, in US dollars.}
  \item{entity5}{recorded values for entity 5, in US dollars.}
  \item{entity6}{recorded values for entity 6, in US dollars.}
  \item{entity7}{recorded values for entity 7, in US dollars.}
  \item{entity8}{recorded values for entity 8, in US dollars.}
  \item{entity9}{recorded values for entity 9, in US dollars.}
  \item{entity10}{recorded values for entity 10, in US dollars.}
  \item{total}{total value, in US dollars.}
}
}
\source{
\url{https://towardsdatascience.com/data-driven-audit-1-automated-
sampling-using-python-52e83347add5}
}
\usage{
data(carrier)
}
\description{
Fictional data from a carrier company in Europe, containing 202 ledger items
across 10 company entities.
}
\examples{
data(carrier)
}
\keyword{datasets}
