% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidenceBound.R
\name{confidenceBound}
\alias{confidenceBound}
\title{Audit confidence / credible bounds}
\usage{
confidenceBound()
}
\value{
A list containing the confidence bound for the audit.
}
\description{
This function
}
\examples{


}
\references{

}
\seealso{

}
\author{
Koen Derks, \email{k.derks@nyenrode.nl}
}
