% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidenceBound.R
\name{confidenceBound}
\alias{confidenceBound}
\title{Audit confidence / credible bounds}
\usage{
confidenceBound(sample = NULL, bookValues = NULL, auditValues = NULL, confidence = 0.95,
                       dataType = "sample", sampleSize = NULL, sumErrors = NULL,
                       method = "binomial", materiality = NULL)
}
\arguments{
\item{sample}{a data frame containing at least a column of book values and a column of audit (true) values.}

\item{bookValues}{the column name for the book values in the sample.}

\item{auditValues}{the column name for the audit (true) values in the sample.}

\item{confidence}{the required confidence level for the bound.}

\item{dataType}{can be either \emph{sample} for data input, or \emph{sumstats} for input in the form of summary statistics.}

\item{sampleSize}{the number of observations in the sample. Only used when \emph{dataType = "sumstats"}.}

\item{sumErrors}{the sum of the errors found in the sample. Only used when \emph{dataType = "sumstats"}.}

\item{method}{can be either one of \emph{binomial, hypergeometric, poisson, stringer}.}

\item{materiality}{if specified, the function also returns the conclusion of the analysis with respect to the materiality. This value must be specified as a fraction of the total value of the population.}
}
\value{
A list containing the confidence bound for the audit.
}
\description{
This function takes a sample data frame or summary statistics about an evaluated audit sample and calculates
a confidence bound accordint to a specified method.
}
\details{
This section lists the available options for the \emph{methods} argument.

\itemize{
 \item{\emph{binomial}         Clopper-Pearson confidence interval for a binomial rate parameter.}
 \item{\emph{stringer}         The Stringer bound (Stringer, 1963).}
 \item{\emph{stringer-meikle}  Stringer bound with Meikle's correction (Meikle, 1972).}
 \item{\emph{stringer-lta}     Stringer bound with LTA correction (Leslie, Teitlebaum, and Anderson, 1979).}
 \item{\emph{stringer-pvz}     Stringer bound with Pap and van Zuijlen's correction.}
 \item{\emph{rohrbach}         Rohrbach's augmented variance bound.}
 \item{\emph{moment}           Modified moment bound.}
}
}
\examples{

# Using the binomial distribution, calculates the upper confidence bound for a materiality of 5\% 
# when 3 mistakes are found in a sample of 234.

# Frequentist planning (n = 234)
jfaRes <- jfa::sampleSize(materiality = 0.05, confidence = 0.95,
                          expectedError = 0.025, likelihood = "binomial")

# Six errors are allowed in the sample. Three are found.

confidenceBound(sampleSize = jfaRes$sampleSize, sumErrors = 3, method = "binomial")

}
\references{
Bickel, P. J. (1992). Inference and auditing: the Stringer bound.
International Statistical Review, 197-209.

Clopper, C. J., & Pearson, E. S. (1934). The use of confidence or
fiducial limits illustrated in the case of the binomial. Biometrika, 26(4),
404-413.

Leslie, D. A., Teitlebaum, A. D., & Anderson, R. J. (1979).
Dollar-unit sampling: a practical guide for auditors. Copp Clark Pitman;
Belmont, Calif.: distributed by Fearon-Pitman.

Meikle, G. R. (1972). Statistical Sampling in an Audit Context:
An Audit Technique. Canadian Institute of Chartered Accountants.

Pap, G., & van Zuijlen, M. C. (1996). On the asymptotic behaviour
of the Stringer bound 1. Statistica Neerlandica, 50(3), 367-389.

Stringer, K. W. (1963). Practical aspects of statistical sampling
in auditing. In Proceedings of the Business and Economic Statistics Section
(pp. 405-411). American Statistical Association.


}
\seealso{

}
\author{
Koen Derks, \email{k.derks@nyenrode.nl}
}
\keyword{bound}
\keyword{confidence}
