% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidenceBound.R
\name{confidenceBound}
\alias{confidenceBound}
\title{Audit confidence / credible bounds}
\usage{
confidenceBound(sample = NULL, bookValues = NULL, auditValues = NULL, confidence = 0.95,
                       dataType = "sample", sampleSize = NULL, sumErrors = NULL,
                       method = "binomial", materiality = NULL)
}
\arguments{
\item{sample}{a data frame containing a column of book values and a column of audit values.}

\item{bookValues}{the column name for the book values in the sample.}

\item{auditValues}{the column name for the audit (true) values in the sample.}

\item{confidence}{the required confidence for the bound.}

\item{dataType}{can be either "sample" for data input, or "sumstats" for input in the form of summary statistics.}

\item{sampleSize}{the number of observations in the sample. Only used when dataType = "sumstats".}

\item{sumErrors}{the sum of the errors found in the sample. Only used when dataType = "sumstats".}

\item{method}{can be either one of "binomial", "hypergeometric", "poisson", "stringer".}

\item{materiality}{if specified, the function also returns the conclusion of the analysis with respect to the materiality. This value be specified as a fraction of the total value of the population.}
}
\value{
A list containing the confidence bound for the audit.
}
\description{
This function takes summary statistics about an evaluated audit sample and calculates
a confidence bound accordint to a specified method.
}
\details{
This section lists the available methods for calculating confidence bounds.

\itemize{
 \item{binomial: The}
 \item{stringer: The Stringer bound is the most well-known and used bound in the accounting
practice. Proposed by Stringer (1963), the bound estimates the mean taint per
 dollar unit.
It's formula is defined as:
\deqn{p(0; 1 - \alpha) + \sum_{j=1}^{m_+} \left[ p(j; 1 - \alpha) -
p(j-1; 1 - \alpha) \right] \cdot z_+_j}{p(0; 1 - \alpha) +
\sum [ p(j; 1 - \alpha) - p(j-1; 1 - \alpha) ] x Z}
where \eqn{p(j; 1 - \alpha)} equals the unique solution to:
\deqn{\sum^n_{k = j + 1} {n \choose k} p^k (1-p)^{n-k} = 1- \alpha}{\sum
(n choose k) p^k (1-p)^{n-k} = 1- \alpha}
and is therefore the Clopper-Pearson confidence interval for a binomial
parameter (Clopper-Pearson, 1934). The values \emph{Z} are the proportional
taints defined as \eqn{\frac{bookValues - auditValues}{bookValues}}
{(bookValues - auditValues) / bookValues}. Since the upper bound is only
defined for integer values of \emph{k}, when partial taints are observed
Stringer performs linear interpolation between the upper bounds that are
properly defined. The Stringer bound is often used for its capacity to yield
sensible results, even when zero errors are found. This comes with a
downside, as the bounds that are given by the Stringer approach are highly
conservative bounds due to the attribute sampling method. Additionally, the
Stringer bound does not accommodate understatements.}
}
}
\examples{

# Using the binomial distribution, calculates the upper confidence bound for a materiality of 5\% 
# when 3 mistakes are found in a sample of 234.

# Frequentist planning (n = 234)
jfaRes <- jfa::sampleSize(materiality = 0.05, confidence = 0.95,
                          expectedError = 0.025, likelihood = "binomial")

# Six errors are allowed in the sample. Three are found.

confidenceBound(sampleSize = jfaRes$sampleSize, sumErrors = 3, method = "binomial")

}
\references{

}
\seealso{

}
\author{
Koen Derks, \email{k.derks@nyenrode.nl}
}
\keyword{bound}
\keyword{confidence}
