% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{evaluation}
\alias{evaluation}
\title{Evaluation of audit samples with confidence / credible bounds}
\usage{
evaluation(sample = NULL, bookValues = NULL, auditValues = NULL, 
                  confidence = 0.95, dataType = "sample", sampleSize = NULL, 
                  sumErrors = NULL, method = "binomial", materiality = NULL, 
                  N = NULL, rohrbachDelta = 2.7)
}
\arguments{
\item{sample}{a data frame containing at least a column of book values and a column of audit (true) values.}

\item{bookValues}{the column name for the book values in the sample.}

\item{auditValues}{the column name for the audit (true) values in the sample.}

\item{confidence}{the required confidence level for the bound.}

\item{dataType}{can be either \emph{sample} for data input, or \emph{sumstats} for input in the form of summary statistics.}

\item{sampleSize}{the number of observations in the sample. Only used when \emph{dataType = "sumstats"}.}

\item{sumErrors}{the sum of the errors found in the sample. Only used when \emph{dataType = "sumstats"}.}

\item{method}{can be either one of \emph{binomial, hypergeometric, poisson, stringer}.}

\item{materiality}{if specified, the function also returns the conclusion of the analysis with respect to the materiality. This value must be specified as a fraction of the total value of the population.}
}
\value{
An object of class \emph{jfaEvaluation}.
}
\description{
This function takes a sample data frame or summary statistics about an evaluated audit sample and calculates a confidence bound accordint to a specified method.
}
\details{
This section lists the available options for the \emph{methods} argument.

\itemize{
 \item{\emph{binomial}         Clopper-Pearson confidence interval for a binomial rate parameter.}
 \item{\emph{stringer}         The Stringer bound (Stringer, 1963).}
 \item{\emph{stringer-meikle}  Stringer bound with Meikle's correction (Meikle, 1972).}
 \item{\emph{stringer-lta}     Stringer bound with LTA correction (Leslie, Teitlebaum, and Anderson, 1979).}
 \item{\emph{stringer-pvz}     Stringer bound with Pap and van Zuijlen's correction (Pap and van Zuijlen, 1996).}
 \item{\emph{rohrbach}         Rohrbach's augmented variance bound (Rohrbach, 1993).}
 \item{\emph{moment}           Modified moment bound (Dworin and Grimlund, 1986).}
}
}
\examples{

library(jfa)

# Generate some audit data (N = 1000)
set.seed(1)
data <- data.frame(ID = sample(1000:100000, size = 1000, replace = FALSE), 
                  bookValue = runif(n = 1000, min = 1000, max = 10000))

# Using the binomial likelihood, calculates the upper 95\% confidence bound for a 
# materiality of 5\% when 1\% full errors are found in a sample (n = 93).
jfaRes <- planning(materiality = 0.05, confidence = 0.95, expectedError = 0.01, 
                   likelihood = "binomial")

# Using monetary unit sampling, draw a random sample from the population.
samp <- sampling(population = data, sampleSize = jfaRes$sampleSize, 
                 units = "mus", bookValues = "bookValue", algorithm = "random")

samp$sample$trueValue <- samp$sample$bookValue
samp$sample$trueValue[2] <- 1561.871 - 500 # One overstatement is found

# Evaluate the sample using the stringer bound.
evaluation(sample = samp$sample, bookValues = "bookValue", 
           auditValues = "trueValue", method = "stringer", materiality = 0.05)

# Evaluate the sample using summary statistics.
evaluation(sampleSize = nrow(samp$sample), sumErrors = 1, dataType = "sumstats",
           method = "binomial", materiality = 0.05)

}
\references{
Bickel, P. J. (1992). Inference and auditing: the Stringer bound.
International Statistical Review, 197-209.

Clopper, C. J., & Pearson, E. S. (1934). The use of confidence or
fiducial limits illustrated in the case of the binomial. Biometrika, 26(4),
404-413.

Dworin, L., & Grimlund, R. A. (1986). Dollar-unit sampling: A
comparison of the quasi-Bayesian and moment bounds. Accounting Review, 36-57.

Leslie, D. A., Teitlebaum, A. D., & Anderson, R. J. (1979).
Dollar-unit sampling: a practical guide for auditors. Copp Clark Pitman;
Belmont, Calif.: distributed by Fearon-Pitman.

Meikle, G. R. (1972). Statistical Sampling in an Audit Context:
An Audit Technique. Canadian Institute of Chartered Accountants.

Pap, G., & van Zuijlen, M. C. (1996). On the asymptotic behaviour
of the Stringer bound 1. Statistica Neerlandica, 50(3), 367-389.

Rohrbach, K. J. (1993). Variance augmentation to achieve nominal
coverage probability in sampling from audit populations. Auditing, 12(2), 79.

Stringer, K. W. (1963). Practical aspects of statistical sampling
in auditing. In Proceedings of the Business and Economic Statistics Section
(pp. 405-411). American Statistical Association.
}
\seealso{

}
\author{
Koen Derks, \email{k.derks@nyenrode.nl}
}
\keyword{bound}
\keyword{confidence}
\keyword{evaluation}
