% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{evaluation}
\alias{evaluation}
\title{Audit Sampling: Evaluation}
\usage{
evaluation(materiality = NULL, min.precision = NULL, method = c(
             "poisson", "binomial", "hypergeometric",
             "stringer", "stringer.meikle", "stringer.lta", "stringer.pvz",
             "rohrbach", "moment", "coxsnell",
             "direct", "difference", "quotient", "regression", "mpu"
           ), alternative = c('less', 'two.sided', 'greater'), conf.level = 0.95,
           data = NULL, values = NULL, values.audit = NULL, strata = NULL, times = NULL,
           x = NULL, n = NULL, N.units = NULL, N.items = NULL,
           pooling = c("none", "complete", "partial"),
           prior = FALSE)
}
\arguments{
\item{materiality}{a numeric value between 0 and 1 specifying the performance materiality (i.e., the maximum tolerable misstatement) as a fraction of the total number of units in the population. Can be \code{NULL}, but \code{min.precision} should be specified in that case. Not used for methods \code{direct}, \code{difference}, \code{quotient}, and \code{regression}.}

\item{min.precision}{a numeric value between 0 and 1 specifying the minimum precision (i.e., upper bound minus most likely error) as a fraction of the total population size. Can be \code{NULL}, but \code{materiality} should be specified in that case.}

\item{method}{a character specifying the inference method. Possible options are \code{poisson} (default), \code{binomial}, \code{hypergeometric}, \code{mpu}, \code{stringer}, \code{stringer.meikle}, \code{stringer.lta}, \code{stringer.pvz}, \code{rohrbach}, \code{moment}, \code{direct}, \code{difference}, \code{quotient}, or \code{regression}. See the details section for more information.}

\item{alternative}{a character indicating the alternative hypothesis and the type of confidence / credible interval. Possible options are \code{less} (default), \code{two.sided}, or \code{greater}.}

\item{conf.level}{a numeric value between 0 and 1 specifying the confidence level.}

\item{data}{a data frame containing a data sample.}

\item{values}{a character specifying name of a numeric column in \code{data} containing the book values of the items.}

\item{values.audit}{a character specifying name of a numeric column in \code{data} containing the audit (true) values of the items.}

\item{strata}{a character specifying name of a factor column in \code{data} indicating to which stratum each item belongs.}

\item{times}{a character specifying name of a numeric column in \code{data} containing the number of times each item in \code{data} should be counted (e.g., due to being selected multiple times for the sample).}

\item{x}{a numeric value larger than 0 specifying the sum of (proportional) misstatements in the sample. If specified, overrides the \code{data}, \code{values} and \code{values.audit} arguments and assumes that the data come from summary statistics specified by both \code{x} and \code{n}.}

\item{n}{an integer larger than 0 specifying the number of items in the sample. If specified, overrides the \code{data}, \code{values} and \code{values.audit} arguments and assumes that the data come from summary statistics specified by both \code{x} and \code{n}.}

\item{N.units}{an integer larger than 0 specifying the number of units in the population. Only used for methods \code{hypergeometric}, \code{direct}, \code{difference}, \code{quotient}, and \code{regression}.}

\item{N.items}{an integer larger than 0 specifying the number of items in the population. Only used for methods \code{direct}, \code{difference}, \code{quotient}, and \code{regression}.}

\item{pooling}{a character specifying the type of model to use when analyzing stratified samples. Possible options are \code{none} (default) for no pooling (i.e., no information is shared between strata), \code{complete} for complete pooling (i.e., all information is shared between strata) or \code{partial} for partial pooling (i.e., information can be shared between strata). The latter option fits a pre-compiled stan model to the data using a MCMC sampling procedure whose options can be set globally using \code{options("mcmc.iterations")} (otherwise: 2000), \code{options("mcmc.warmup")} (otherwise: 1000), \code{options("mcmc.chains")} (otherwise: 4) and \code{options("mcmc.cores")} (otherwise: 1).}

\item{prior}{a logical specifying whether to use a prior distribution, or an object of class \code{jfaPrior} or \code{jfaPosterior} containing the prior distribution. If \code{FALSE} (default), performs classical planning. If \code{TRUE}, performs Bayesian planning using a default conjugate prior.}
}
\value{
An object of class \code{jfaEvaluation} containing:

\item{conf.level}{a numeric value between 0 and 1 giving the confidence level.}
\item{mle}{a numeric value between 0 and 1 giving the most likely error in the population.}
\item{ub}{a numeric value between 0 and 1 giving the upper bound for the population misstatement.}
\item{lb}{{a numeric value between 0 and 1 giving the lower bound for the population misstatement.}}
\item{precision}{a numeric value between 0 and 1 giving the difference between the most likely error and the upper bound.}
\item{p.value}{for classical tests, a numeric value giving the one-sided p-value.}
\item{x}{an integer larger than, or equal to, 0 giving the number of sample errors.}
\item{t}{a value larger than, or equal to, 0, giving the sum of proportional sample errors.}
\item{n}{an integer larger than 0 giving the sample size.}
\item{materiality}{if \code{materiality} is specified, a numeric value between 0 and 1 giving the performance materiality as a fraction of the number of units in the population.}
\item{min.precision}{if \code{min.precision} is specified, a numeric value between 0 and 1 giving the minimum precision as a fraction of the number of units in the population.}
\item{alternative}{a character indicating the alternative hypothesis.}
\item{method}{a character indicating the inference method.}
\item{N.units}{if \code{N.units} is specified, in integer larger than 0 indicating the number of units in the population.}
\item{N.items}{if \code{N.items} is specified, in integer larger than 0 indicating the number of items in the population.}
\item{K}{if \code{method = 'hypergeometric'}, an integer indicating the assumed total errors in the population.}
\item{prior}{an object of class 'jfaPrior' that contains the prior distribution.}
\item{posterior}{an object of class 'jfaPosterior' that contains the posterior distribution.}
\item{data}{a data frame containing the relevant columns from the \code{data}.}
\item{strata}{a data frame containing the relevant statistical results for the strata.}
\item{data.name}{a character giving the name of the data.}
}
\description{
\code{evaluation()} is used to perform statistical inference about the misstatement in an audit population. It allows specification of statistical requirements for the sample with respect to the performance materiality or the precision. The function returns an object of class \code{jfaEvaluation} that can be used with associated \code{summary()} and \code{plot()} methods.
}
\details{
This section lists the available options for the \code{methods} argument.

\itemize{
 \item{\code{poisson}:         Evaluates the sample with the Poisson distribution. If combined with \code{prior = TRUE}, performs Bayesian evaluation using a \emph{gamma} prior and posterior.}
 \item{\code{binomial}:        Evaluates the sample with the binomial distribution. If combined with \code{prior = TRUE}, performs Bayesian evaluation using a \emph{beta} prior and posterior.}
 \item{\code{hypergeometric}:  Evaluates the sample with the hypergeometric distribution. If combined with \code{prior = TRUE}, performs Bayesian evaluation using a \emph{beta-binomial} prior and posterior.}
	\item{\code{mpu}:             Evaluates the sample with the mean-per-unit estimator.}
 \item{\code{stringer}:        Evaluates the sample with the Stringer bound (Stringer, 1963).}
 \item{\code{stringer.meikle}: Evaluates the sample with the Stringer bound with Meikle's correction for understatements (Meikle, 1972).}
 \item{\code{stringer.lta}:    Evaluates the sample with the Stringer bound with LTA correction for understatements (Leslie, Teitlebaum, and Anderson, 1979).}
 \item{\code{stringer.pvz}:    Evaluates the sample with the Stringer bound with Pap and van Zuijlen's correction for understatements (Pap and van Zuijlen, 1996).}
 \item{\code{rohrbach}:        Evaluates the sample with Rohrbach's augmented variance bound (Rohrbach, 1993).}
 \item{\code{moment}:          Evaluates the sample with the modified moment bound (Dworin and Grimlund, 1984).}
 \item{\code{coxsnell}:        Evaluates the sample with the Cox and Snell bound (Cox and Snell, 1979).}
 \item{\code{direct}:          Evaluates the sample with the direct estimator (Touw and Hoogduin, 2011).}
 \item{\code{difference}:      Evaluates the sample with the difference estimator (Touw and Hoogduin, 2011).}
 \item{\code{quotient}:        Evaluates the sample with the quotient estimator (Touw and Hoogduin, 2011).}
 \item{\code{regression}:      Evaluates the sample with the regression estimator (Touw and Hoogduin, 2011).}
}
}
\examples{
# Using summary statistics
evaluation(materiality = 0.05, x = 0, n = 100)
evaluation(materiality = 0.05, x = c(2, 1, 0), n = c(50, 70, 40))

# Using data
data("BuildIt")
BuildIt$inSample <- c(rep(1, 100), rep(0, 3400))
levs <- c("low", "medium", "high")
BuildIt$stratum <- factor(c(levs[3], levs[2], rep(levs, times = 1166)))
sample <- subset(BuildIt, BuildIt$inSample == 1)

evaluation(
  materiality = 0.05, data = sample,
  values = "bookValue", values.audit = "auditValue"
)
evaluation(
  materiality = 0.05, data = sample, values = "bookValue",
  values.audit = "auditValue", strata = "stratum"
)
}
\references{
Cox, D. and Snell, E. (1979). On sampling and the estimation of rare errors. \emph{Biometrika}, 66(1), 125-132.

Derks, K., de Swart, J., van Batenburg, P., Wagenmakers, E.-J., & Wetzels, R. (2021). Priors in a Bayesian audit: How integration of existing information into the prior distribution can improve audit transparency and efficiency. \emph{International Journal of Auditing}, 25(3), 621-636.

Dworin, L. D. and Grimlund, R. A. (1984). Dollar-unit sampling for accounts receivable and inventory. \emph{The Accounting Review}, 59(2), 218-241

Leslie, D. A., Teitlebaum, A. D., & Anderson, R. J. (1979). \emph{Dollar-unit Sampling: A Practical Guide for Auditors}. Copp Clark Pitman; Belmont, Calif.: distributed by Fearon-Pitman.

Meikle, G. R. (1972). \emph{Statistical Sampling in an Audit Context: An Audit Technique}. Canadian Institute of Chartered Accountants.

Pap, G., and van Zuijlen, M. C. (1996). On the asymptotic behavior of the Stringer bound. \emph{Statistica Neerlandica}, 50(3), 367-389.

Rohrbach, K. J. (1993). Variance augmentation to achieve nominal coverage probability in sampling from audit populations. \emph{Auditing}, 12(2), 79.

Stringer, K. W. (1963). Practical aspects of statistical sampling in auditing. \emph{In Proceedings of the Business and Economic Statistics Section} (pp. 405-411). American Statistical Association.

Touw, P., and Hoogduin, L. (2011). \emph{Statistiek voor Audit en Controlling}. Boom uitgevers Amsterdam.
}
\seealso{
\code{\link{auditPrior}} \code{\link{planning}} \code{\link{selection}} \code{\link{report}}
}
\author{
Koen Derks, \email{k.derks@nyenrode.nl}
}
\keyword{audit}
\keyword{evaluation}
\keyword{prior}
