% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jfa-convolution.R
\name{jfa-convolution}
\alias{jfa-convolution}
\alias{+.jfaPrior}
\alias{+.jfaPosterior}
\alias{*.jfaPrior}
\alias{*.jfaPosterior}
\title{Prior convolution utility functions}
\usage{
\method{+}{jfaPrior}(p1, p2)

\method{+}{jfaPosterior}(p1, p2)

\method{*}{jfaPrior}(c, p1)

\method{*}{jfaPosterior}(c, p1)
}
\arguments{
\item{p1}{an object of class \code{jfaPrior}.}

\item{p2}{an object of class \code{jfaPrior}.}

\item{c}{a numeric value, used for multiplication.}
}
\value{
An object of class \code{jfaPrior} with method \code{convolution}.
}
\description{
Methods defined for convolutions of \code{jfaPrior} objects.
}
