% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planning.R
\name{planning}
\alias{planning}
\title{Frequentist and Bayesian Planning for Audit Samples}
\usage{
planning(materiality, confidence = 0.95, expectedError = 0, likelihood = "poisson", 
         minPrecision = NULL, N = NULL, maxSize = 5000, increase = 1, prior = FALSE, kPrior = 0, nPrior = 0)
}
\arguments{
\item{materiality}{a value between 0 and 1 representing the materiality of the audit as a fraction of the total size or value. Can be \code{NULL}, but \code{minPrecision} should be specified in that case.}

\item{confidence}{the confidence level desired from the confidence bound (on a scale from 0 to 1). Defaults to 0.95, or 95\% confidence.}

\item{expectedError}{a fraction representing the percentage of expected mistakes in the sample relative to the total size, or a number (>= 1) that represents the number of expected mistakes.}

\item{minPrecision}{The minimum precision to be obtained. Can be \code{NULL}, but \code{materiality} should be specified in that case.}

\item{likelihood}{can be one of \code{binomial}, \code{poisson}, or \code{hypergeometric}.}

\item{N}{the population size (required for hypergeometric calculations).}

\item{maxSize}{the maximum sample size that is considered for calculations. Defaults to 5000 for efficiency. Increase this value if the sample size cannot be found due to it being too large (e.g., for a low materiality).}

\item{increase}{the desired increase step for the sample size calculation.}

\item{prior}{whether to use a prior distribution when planning. Defaults to \code{FALSE} for frequentist planning. If \code{TRUE}, the prior distribution is updated by the specified likelihood. Chooses a conjugate gamma distribution for the Poisson likelihood, a conjugate beta distribution for the binomial likelihood, and a conjugate beta-binomial distribution for the hypergeometric likelihood.}

\item{kPrior}{the prior parameter \eqn{\alpha} (number of errors in the assumed prior sample).}

\item{nPrior}{the prior parameter \eqn{\beta} (total number of observations in the assumed prior sample).}
}
\value{
An object of class \code{jfaPlanning} containing:

\item{materiality}{the value of the specified materiality. Can be \code{NULL}.}
\item{confidence}{the confidence level for the desired population statement.}
\item{sampleSize}{the resulting sample size.}
\item{expectedSampleError}{the number of full errors that are allowed to occur in the sample.}
\item{expectedError}{the specified number of errors as a fraction or as a number.}
\item{likelihood}{the specified likelihood.}
\item{errorType}{whether the expected errors where specified as a percentage or as an integer.}
\item{minPrecision}{The minimum precision to be obtained. Can be \code{NULL}.}
\item{N}{the population size (only returned in case of a hypergeometric likelihood).}
\item{populationK}{the assumed population errors (only returned in case of a hypergeometric likelihood).}
\item{prior}{a list containing information on the prior parameters.}
}
\description{
This function calculates the required sample size for an audit, based on the poisson, binomial, or hypergeometric likelihood. A prior can be specified to perform Bayesian planning. The returned object is of class \code{jfaPlanning} and can be used with associated \code{print()} and \code{plot()} methods.
}
\details{
This section elaborates on the available likelihoods and corresponding prior distributions for the \code{likelihood} argument.

\itemize{
 \item{\code{poisson}:          The Poisson likelihood is used as a likelihood for monetary unit sampling (MUS). Its likelihood function is defined as: \deqn{p(x) = \frac{\lambda^x e^{-\lambda}}{x!}} The conjugate \emph{gamma(\eqn{\alpha, \beta})} prior has probability density function: \deqn{f(x; \alpha, \beta) = \frac{\beta^\alpha x^{\alpha - 1} e^{-\beta x}}{\Gamma(\alpha)}}}
 \item{\code{binomial}:         The binomial likelihood is used as a likelihood for record sampling \emph{with} replacement. Its likelihood function is defined as: \deqn{p(x) = {n \choose k} p^k (1 - p)^{n - k}} The conjugate \emph{beta(\eqn{\alpha, \beta})} prior has probability density function: \deqn{f(x; \alpha, \beta) = \frac{1}{B(\alpha, \beta)} x^{\alpha - 1} (1 - x)^{\beta - 1}}}
 \item{\code{hypergeometric}:   The hypergeometric likelihood is used as a likelihood for record sampling \emph{without} replacement. Its likelihood function is defined as: \deqn{p(x = k) = \frac{{K \choose k} {N - K \choose n - k}}{{N \choose n}}} The conjugate \emph{beta-binomial(\eqn{\alpha, \beta})} prior (Dyer and Pierce, 1993) has probability density function: \deqn{f(k | n, \alpha, \beta) = {n \choose k} \frac{B(k + \alpha, n - k + \beta)}{B(\alpha, \beta)}} }
}
}
\examples{
library(jfa)

# Using the binomial distribution, calculates the required sample size for a 
# materiality of 5\% when 2.5\% mistakes are expected to be found in the sample.

# Frequentist planning with binomial likelihood:

p1 <- planning(materiality = 0.05, confidence = 0.95, expectedError = 0.025, 
               likelihood = "binomial")
print(p1)

# jfa planning results for binomial likelihood:
#      
# Materiality:             5\% 
# Confidence:              95\% 
# Sample size:             234 
# Allowed sample errors:   6

# Bayesian planning with uninformed prior:

p2 <- planning(materiality = 0.05, confidence = 0.95, expectedError = 0.025, 
               likelihood = "binomial", prior = TRUE)
print(p2)

# jfa planning results for beta prior with binomial likelihood:
#      
# Materiality:             5\% 
# Confidence:              95\% 
# Sample size:             220 
# Allowed sample errors:   5.5 
# Prior parameter alpha:   1 
# Prior parameter beta:    1

# Bayesian planning with informed prior:

prior <- auditPrior(materiality = 0.05, confidence = 0.95, cr = 0.6, 
                    expectedError = 0.025, likelihood = "binomial")

p3 <- planning(materiality = 0.05, confidence = 0.95, expectedError = 0.025,
               prior = prior)
print(p3)

# jfa planning results for beta prior with binomial likelihood:
#      
# Materiality:             5\% 
# Confidence:              95\% 
# Sample size:             169 
# Allowed sample errors:   4.23 
# Prior parameter alpha:   2.275 
# Prior parameter beta:    50.725

}
\references{
Dyer, D. and Pierce, R.L. (1993). On the Choice of the Prior Distribution in Hypergeometric Sampling. \emph{Communications in Statistics - Theory and Methods}, 22(8), 2125 - 2146.
}
\seealso{
\code{\link{auditPrior}} \code{\link{sampling}} \code{\link{evaluation}}
}
\author{
Koen Derks, \email{k.derks@nyenrode.nl}
}
\keyword{audit}
\keyword{planning}
\keyword{sample}
\keyword{size}
