% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planning.R
\name{planning}
\alias{planning}
\title{Frequentist and Bayesian planning for audit samples}
\usage{
planning(materiality, confidence = 0.95, expectedError = 0, 
                likelihood = "poisson", N = NULL, maxSize = 5000, 
                prior = FALSE, kPrior = 0, nPrior = 0)
}
\arguments{
\item{materiality}{a value between 0 and 1 representing the materiality of the audit as a fraction of the total size or value.}

\item{confidence}{the confidence level desired from the confidence bound (on a scale from 0 to 1). Defaults to 0.95, or 95\% confidence.}

\item{expectedError}{a fraction representing the percentage of expected mistakes in the sample relative to the total size, or a number (> 1) that represents the number of expected mistakes.}

\item{likelihood}{can be one of \emph{binomial}, \emph{poisson}, or \emph{hypergeometric}.}

\item{N}{the population size (required for hypergeometric calculations).}

\item{maxSize}{the maximum sample size that is considered for calculations. Increase this value if the samle size cannot be found due to it being too large (e.g., for low materialities).}

\item{prior}{If TRUE, adds a prior distribution to be updated by the specified likelihood. Chooses a conjugate beta distribution for the binomial likelihood, a conjugate gamma distribution for the poisson likelihood, and a conjugate beta-binomial distribution for the hypergeometric likelihood. Defaults to FALSE for frequentist planning.}

\item{kPrior}{the prior parameter \eqn{\alpha} (errors in the assumed prior sample).}

\item{nPrior}{the prior parameter \eqn{\beta} (sample size of assumed prior sample).}
}
\value{
An object of class \emph{jfaPlanning}.
}
\description{
This function calculates the required sample size for an audit, based on the poisson, binomial, or hypergeometric likelihood. A prior can be specified to perform Bayesian planning. The returned object has a print() and a plot() method.
}
\examples{

library(jfa)

# Using the binomial distribution, calculates the required sample size for a 
# materiality of 5\% when 2.5\% mistakes are expected to be found in the sample.

# Frequentist planning with binomial likelihood.
planning(materiality = 0.05, confidence = 0.95, expectedError = 0.025, 
         likelihood = "binomial")

# Bayesian planning with uninformed prior.
planning(materiality = 0.05, confidence = 0.95, expectedError = 0.025, 
         likelihood = "binomial", prior = TRUE)

# Bayesian planning with informed prior (based on 10 correct observations).
planning(materiality = 0.05, confidence = 0.95, expectedError = 0.025, 
         likelihood = "binomial", prior = TRUE, kPrior = 0, nPrior = 10)

}
\references{

}
\seealso{
\code{\link{sampling}} \code{\link{evaluation}}
}
\author{
Koen Derks, \email{k.derks@nyenrode.nl}
}
\keyword{planning}
\keyword{sample}
\keyword{size}
