% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retailer-data.R
\docType{data}
\name{retailer}
\alias{retailer}
\title{Retailer data}
\format{
A data frame with 20 rows and 5 variables.
\describe{
  \item{i}{branch number.}
  \item{N}{total number of items in each branch.}
  \item{n}{number of items in sample per branch.}
  \item{x}{number of errors in sample per branch.}
}
}
\usage{
data(retailer)
}
\description{
Sample outcomes per stratum from a retail company consisting of 20 branches.
}
\examples{
data(retailer)
}
\references{
Derks, K., de Swart, J., & Wetzels, R. (2022). Een Bayesiaanse blik op gestratificeerde steekproeven heeft voordelen voor de auditor. \emph{Maandblad voor Accountancy en Bedrijfseconomie}, 96(1/2), 37-46. https://doi.org/10.5117/mab.96.78836
}
\keyword{datasets}
